requirements.txt
mulet-nbt>=2.1.1,<3
numpy>=1.19.0,<1.20
